+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.2.2
 
+* `Utils`:
+  * `buildEntities` now is inline
 * `Behaviour Builder`:
   * Fixes in `CallbackQuery` triggers